+++ libs/common/src/lib/types/access-level.type.ts
@@ -1,0 +1,2 @@
+export type AccessLevel =
+  'CREATE_READ_UPDATE_DELETE' | 'READ' | 'READ_RESTRICTED';