+++ libs/ui/src/lib/access-level-icon/access-level-icon.component.stories.ts
@@ -1,0 +1,42 @@
+import { CommonModule } from '@angular/common';
+import { IonIcon } from '@ionic/angular/standalone';
+import { moduleMetadata } from '@storybook/angular';
+import type { Meta, StoryObj } from '@storybook/angular';
+
+import { GfAccessLevelIconComponent } from './access-level-icon.component';
+
+export default {
+  title: 'Access Level Icon',
+  component: GfAccessLevelIconComponent,
+  decorators: [
+    moduleMetadata({
+      imports: [CommonModule, IonIcon]
+    })
+  ],
+  argTypes: {
+    accessLevel: {
+      control: 'select',
+      options: ['CREATE_READ_UPDATE_DELETE', 'READ', 'READ_RESTRICTED']
+    }
+  }
+} as Meta<GfAccessLevelIconComponent>;
+
+type Story = StoryObj<GfAccessLevelIconComponent>;
+
+export const RestrictedView: Story = {
+  args: {
+    accessLevel: 'READ_RESTRICTED'
+  }
+};
+
+export const View: Story = {
+  args: {
+    accessLevel: 'READ'
+  }
+};
+
+export const ViewAndManage: Story = {
+  args: {
+    accessLevel: 'CREATE_READ_UPDATE_DELETE'
+  }
+};